requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
bikram==2.1.0
CacheControl==0.12.6
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.3
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.1
cssselect2==0.3.0
daphne==2.5.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.3
django-cleanup==5.0.0
django-cors-headers==3.2.1
django-extensions==2.2.8
django-heroku==0.3.1
django-jwt-auth==0.0.2
django-user-activity-log2==0.0.24
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Faker==4.0.2
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==1.0.0
numpy==1.19.1
packaging==20.3
pandas==1.1.0
pep517==0.8.2
Pillow==7.1.1
progress==1.5
psutil==5.7.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
Pyphen==0.9.5
python-barcode==0.11.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
qrcode==6.1
redis==3.5.3
redislite==5.0.165407
reportlab==3.5.53
requests==2.23.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
tinycss2==1.0.2
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
uWSGI==2.0.19.1
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
xlwt==1.3.0
zope.interface==5.1.2
